Limit is_uri check to plausible schemes
The is_uri function checked whether a ":" appeared before a "/" to try
to detect a URI. This had too many false positives as reported in #44,
so this commit limits the test to syntactically-valid URI schemes - if
any non-scheme byte appears before the colon, it will be resolved as a
non-URI.

There can still be ambiguous cases, in which case using "./filename" is
a suitable workaround. URI remains the default when the argument can be
interpreted as beginning with a valid scheme.

Closes #44