fix argument ambiguity by using double-dash
since it's valid for filenames to begin with a dash, there's an
ambiguity on weather something like `-i` refers to the cli flag or a
file named `-i`.

the convention is to use a double-dash "--" as a way to mark the end of
cli options so that everything that comes after it can be treated as
arguments.

Closes: https://github.com/mwh/dragon/issues/59