Use relative path as default, switch with -f/-F
By default, a button will display the relative path to a file within
the current working directory, if it is in there, and otherwise
will display an absolute path.

With the -f/--name-only option, only the basename will be displayed
for any file.

With the -F/--full-path option, only absolute paths will be used for
any file; this was the default previously.